Cant get FOV to work
I think that the problem are that are walls are not GameObjects. I will leave fov at side for a while and will concentrate today efforts in finishing more things.

+++ include/map.h
@@ -27,7 +27,7 @@ typedef struct Segment {
 // slightly differently
 typedef struct Wall {
 
-    u8 x, y; // position
+    u32 x, y; // position
 
     // graphics
     // asciiChar glyph;
